requirements.txt
pyaarlo
azure.storage.blob
sqlalchemy
pyodbc
python-dotenv
image
Flask